third_party/blink/web_tests/external/wpt/cors-rfc1918/address-space.window-expected.txt

This is a testharness.js-based test.
PASS About:blank iframe's addressSpace is inherited from the root.
PASS About:srcdoc iframe's addressSpace is inherited from the root.
PASS Data: iframe's addressSpace is inherited from the root.
PASS Blob: iframe's addressSpace is inherited from the root.
PASS Local iframe's addressSpace is local.
PASS Treat-as-public-address iframe's addressSpace is public.
FAIL Local-local grandchild iframe's addressSpace is local. assert_equals: grandchild expected "Loaded" but got ""
Harness: the test ran to completion.
